requirements.txt
flask
werkzeug
flask-cors
pdfplumber
python-docx
google-generativeai
python-dotenv
protobuf>=5.26.1
